move trash related types to types/trash
